get_name now uses a toolhelp snapshot instead of reading EnumProcessModules, which is cleaner code and doesn't fail for non-privileged users like EnumProcessModules. It may be a little slower but better slow and return data than fast and fail to return data.
Test suite had to be modified as well, since the previous method of looping EnumProcessModules until ERROR_PARTIAL_COPY completed was actually preventing a race condition from appearing. Now that we're not doing that, race condition re-emerged in test_cmdline and test_path. Edited test suite to wait one more iteration of time.sleep() *after* the PID appears in the process list, so that it has time to fully initialize and allow process memory to be read.



git-svn-id: http://psutil.googlecode.com/svn/trunk@205 d8a29897-7e4b-0410-a74a-8bb4c295346f